Refactor User class to include birthOfDay property
Update the User class to include a new property called birthOfDay. This property represents the birth date of the user. The property is nullable and will be used to store the user's birth date information.

This refactor improves the User class by adding a new property to capture important user data.