feat: Add confirm asset and update gender asset widget
Add a new asset called "confirm" to the asset path constants. This asset will be used for displaying a confirmation icon in various parts of the application.

Update the `GenderAsset` widget to wrap the `Lottie.asset` widget with a `RepaintBoundary` widget. This change improves the performance and stability of the widget.

These changes enhance the application by adding a new asset and improving the gender asset widget.